+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Export `Filaments.reparametrise_arclength!`, which recompues the filament parametrisation
+  in order to approximate the natural (arc length) parametrisation.
+
 ## [0.34.3] - 2026-05-28
 
 ### Changed